Finished main challenge.
