+++ Semana-8/Aula_35_Astro_Match/astro-match/src/actions/matches.js
@@ -1,0 +1,20 @@
+import axios from 'axios';
+
+export const showMatches = (matches) => ({
+	type: "SHOW_MATCHES",
+	payload: {
+		matches,
+	}
+});
+
+export const requestMatches = () => async (dispatch) => {
+	await axios.get(
+		"https://us-central1-missao-newton.cloudfunctions.net/astroMatch/davi/matches"
+	).then((res) => {
+		dispatch(showMatches(
+			res.data.matches
+		))
+	}).catch((err) => {
+        console.log(err)
+    })
+};
