+++ Semana-8/Aula_35_Astro_Match/astro-match/src/containers/SwipeScreen/index.js
@@ -82,7 +82,7 @@ SwipeScreen.propTypes = {
 }
 
 const mapStateToProps = (state) => ({
-	profileToSwipe: state.profiles
+	profileToSwipe: state.profiles,
 })
 
 const mapDispatchToProps = (dispatch) => {
