+++ Semana-8/Aula_35_Astro_Match/astro-match/src/reducers/index.js
@@ -1,10 +1,12 @@
 import { combineReducers } from 'redux'
-import routes from './routes';
+import routes from './routes'
 import profiles from './profiles'
+import matches from './matches'
 
 const rootReducer = combineReducers({
   routes, // equivalente a routes: routes
   profiles, // equivalente a profiles: profiles
+  matches,
 })
 
 export default rootReducer;
