+++ Semana-8/Aula_35_Astro_Match/astro-match/src/reducers/matches.js
@@ -1,0 +1,13 @@
+const initialState = []
+
+const matches = (state = initialState, action) => {
+
+    switch (action.type) {
+      case "SHOW_MATCHES":
+        return Array.from(action.payload.matches);
+      default:
+        return state
+    }
+  }
+  
+  export default matches
